pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "think-drop"
version = "0.1.7"
description = "A Python package for building Telegram bots with ease and efficiency."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.84.0",
    "custom-python-logger>=2.0.13",
    "google-genai>=1.65.0",
    "notion-client>=3.0.0",
    "openai>=2.24.0",
    "pydantic-settings>=2.13.1",
    "pytest>=9.0.2",
    "pytest-plugins>=2.0.2",
    "python-base-toolkit>=1.0.4",
    "python-notion-plus>=1.0.0",
    "python-telegram-bot>=22.6",
    "pytz>=2025.2",
    "tenacity>=9.1.4",
]
keywords = ["cli", "command", "telegram", "bot", "integration", "python"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
]

[project.urls]
Homepage = "https://github.com/aviz92/think-drop"
Repository = "https://github.com/aviz92/think-drop"
Issues = "https://github.com/aviz92/think-drop/issues"

[tool.hatch.build.targets.wheel]
packages = ["think_drop*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 120
